Finalizacion de Archivos Controllers

+++ Entities/Cargo.cs
@@ -5,7 +5,7 @@ namespace WebApiKalum.Entities
         public string CargoId { get; set; }
         public string Descripcion { get; set; }
         public string Prefijo { get; set; }
-        public double Monto { get; set; }
+        public Decimal Monto { get; set; }
         public bool GeneraMora { get; set; }
         public int PorcentajeMora { get; set; }
         public virtual List<CuentaPorCobrar> CuentasPorCobrar { get; set; }
