+++ Entities/InscripcionPago.cs
@@ -6,7 +6,7 @@ namespace WebApiKalum.Entities
         public string Anio { get; set;}
         public string NoExpediente { get; set; }
         public DateTime FechaPago { get; set; }
-        public double Monto { get; set; }
+        public Decimal Monto { get; set; }
         public virtual Aspirante aspirantes { get; set; }
 
     }
