+++ Entities/InversionCarreraTecnica.cs
@@ -4,9 +4,9 @@ namespace WebApiKalum.Entities
     {
         public string InversionId { get; set; }
         public string CarreraId { get; set; }
-        public double MontoInscripcion { get; set; }
+        public Decimal MontoInscripcion { get; set; }
         public int NumeroPagos { get; set;}
-        public double MontoPagos { get; set; }
+        public Decimal MontoPagos { get; set; }
         public virtual CarreraTecnica CarreraTecnica { get; set; }
     }
 }
